fix: Remove zenoh dependency on libzenohc
The dependency on libzenohc makes testing the zenoh formula
quite complicated as it relies on a formulae built outside the
eclipse-zenoh/zenoh repository.

At the same time, it is odd how zenoh includes libzenohc although it
is unused by the router. There is a better argument for including some
of the plugins and backends.